modularised the code :love_letter:
